redis, 필터 내 예외처리, 응답 통일
