--- src/main/java/WeGoTogether/wegotogether/ApiPayload/code/exception/Handler/UserHandler.java
@@ -1,11 +1,0 @@
-package WeGoTogether.wegotogether.ApiPayload.code.exception.Handler;
-
-import WeGoTogether.wegotogether.ApiPayload.code.BaseErrorCode;
-import WeGoTogether.wegotogether.ApiPayload.code.exception.CustomException;
-
-public class UserHandler extends CustomException {
-
-    public UserHandler(BaseErrorCode errorCode) {
-        super(errorCode);
-    }
-}
