+++ src/main/java/WeGoTogether/wegotogether/constant/Handler/UserHandler.java
@@ -1,0 +1,11 @@
+package WeGoTogether.wegotogether.constant.Handler;
+
+import WeGoTogether.wegotogether.constant.code.BaseErrorCode;
+import WeGoTogether.wegotogether.constant.exception.CustomException;
+
+public class UserHandler extends CustomException {
+
+    public UserHandler(BaseErrorCode errorCode) {
+        super(errorCode);
+    }
+}
