+++ src/main/java/WeGoTogether/wegotogether/constant/code/BaseCode.java
@@ -1,4 +1,4 @@
-package WeGoTogether.wegotogether.ApiPayload.code;
+package WeGoTogether.wegotogether.constant.code;
 
 public interface BaseCode {
     ReasonDTO getReason();
