+++ src/main/java/WeGoTogether/wegotogether/constant/code/BaseErrorCode.java
@@ -1,4 +1,4 @@
-package WeGoTogether.wegotogether.ApiPayload.code;
+package WeGoTogether.wegotogether.constant.code;
 
 public interface BaseErrorCode {
     ErrorReasonDTO getReason();
