+++ src/main/java/WeGoTogether/wegotogether/constant/code/ErrorReasonDTO.java
@@ -1,15 +1,16 @@
-package WeGoTogether.wegotogether.ApiPayload.code;
+package WeGoTogether.wegotogether.constant.code;
 
 import lombok.Builder;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
-@Getter
 @Builder
-public class ReasonDTO {
+@Getter
+public class ErrorReasonDTO {
     private HttpStatus httpStatus;
 
-    private final boolean isSuccess;
+
+    private final Boolean isSuccess;
     private final String code;
     private final String message;
