+++ src/main/java/WeGoTogether/wegotogether/constant/code/ReasonDTO.java
@@ -1,16 +1,15 @@
-package WeGoTogether.wegotogether.ApiPayload.code;
+package WeGoTogether.wegotogether.constant.code;
 
 import lombok.Builder;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
-@Builder
 @Getter
-public class ErrorReasonDTO {
+@Builder
+public class ReasonDTO {
     private HttpStatus httpStatus;
 
-
-    private final Boolean isSuccess;
+    private final boolean isSuccess;
     private final String code;
     private final String message;
