+++ src/main/java/WeGoTogether/wegotogether/constant/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package WeGoTogether.wegotogether.domain.common;
+package WeGoTogether.wegotogether.constant.entity;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
@@ -8,8 +8,6 @@ import org.hibernate.annotations.DynamicUpdate;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-
 
 import java.time.LocalDateTime;
