+++ src/main/java/WeGoTogether/wegotogether/security/ExceptionHandlerFilter.java
@@ -1,6 +1,6 @@
 package WeGoTogether.wegotogether.security;
 
-import WeGoTogether.wegotogether.ApiPayload.code.status.ErrorStatus;
+import WeGoTogether.wegotogether.constant.enums.ErrorStatus;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
 import jakarta.servlet.FilterChain;
@@ -8,12 +8,10 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
-import java.util.NoSuchElementException;
 
 @Slf4j
