+++ src/main/java/WeGoTogether/wegotogether/security/SecurityConfig.java
@@ -1,6 +1,6 @@
 package WeGoTogether.wegotogether.security;
 
-import WeGoTogether.wegotogether.security.handler.CustomAccessDeniedHandler;
+import WeGoTogether.wegotogether.constant.Handler.CustomAccessDeniedHandler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Bean;
